refactor: 머지 컨플릭트 해결
- details 브랜치와 admin-page 브랜치간 머지 컨플릭트 해결
