+++ backend/src/main/java/com/jujeol/commons/exception/ExceptionCodeAndDetails.java
@@ -1,9 +1,15 @@
 package com.jujeol.commons.exception;
 
+import com.jujeol.drink.exception.CreateReviewLimitException;
 import com.jujeol.drink.exception.InvalidAlcoholByVolumeException;
 import com.jujeol.drink.exception.InvalidDrinkNameException;
 import com.jujeol.drink.exception.InvalidEnglishNameException;
+import com.jujeol.drink.exception.InvalidReviewContentException;
 import com.jujeol.drink.exception.NotExistReviewInDrinkException;
+import com.jujeol.drink.exception.NotFoundCategoryException;
+import com.jujeol.drink.exception.NotFoundDrinkException;
+import com.jujeol.drink.exception.NotFoundReviewException;
+import com.jujeol.member.exception.InvalidTokenException;
 import com.jujeol.member.exception.KakaoAccessException;
 import com.jujeol.member.exception.NoSuchMemberException;
 import com.jujeol.member.exception.UnauthorizedUserException;
@@ -32,9 +38,9 @@ public enum ExceptionCodeAndDetails {
     NOT_FOUND_REVIEW("2005", "해당하는 id의 리뷰를 찾을 수 없습니다.", NotFoundReviewException.class),
     NOT_EXIST_REVIEW_IN_DRINK("2006", "해당 주류에 존재하지 않는 리뷰입니다.",
             NotExistReviewInDrinkException.class),
-    NOT_EXIST_CATEGORY("2007", "해당 카테고리가 존재하지 않습니다.", NotFoundCategoryException.class),
-    INVALID_CONTENT_LENGTH("2008", "리뷰는 빈 공백이거나 300자를 넘을 수 없습니다.", InvalidReviewContentException.class)
-    ;
+    CREATE_REVIEW_LIMIT	("2007",	"동일 상품에 대한 리뷰는 하루에 하나만 작성할 수 있습니다.", CreateReviewLimitException.class),
+    INVALID_CONTENT_LENGTH("2008", "리뷰는 빈 공백이거나 300자를 넘을 수 없습니다.", InvalidReviewContentException.class),
+    NOT_EXIST_CATEGORY("2009", "해당 카테고리가 존재하지 않습니다.", NotFoundCategoryException.class);
 
     private String code;
     private String message;
