+++ backend/src/main/java/com/jujeol/drink/application/DrinkService.java
@@ -4,7 +4,6 @@ import com.jujeol.drink.application.dto.DrinkDto;
 import com.jujeol.drink.application.dto.DrinkRequestDto;
 import com.jujeol.drink.domain.Drink;
 import com.jujeol.drink.domain.repository.DrinkRepository;
-import com.jujeol.drink.domain.repository.ReviewRepository;
 import com.jujeol.drink.exception.NotFoundDrinkException;
 import com.jujeol.member.domain.Member;
 import com.jujeol.member.domain.Preference;
