+++ backend/src/main/java/com/jujeol/drink/application/ReviewService.java
@@ -4,9 +4,9 @@ import com.jujeol.drink.application.dto.MemberDto;
 import com.jujeol.drink.application.dto.ReviewRequest;
 import com.jujeol.drink.application.dto.ReviewResponse;
 import com.jujeol.drink.domain.Drink;
-import com.jujeol.drink.domain.DrinkRepository;
 import com.jujeol.drink.domain.Review;
-import com.jujeol.drink.domain.ReviewRepository;
+import com.jujeol.drink.domain.repository.DrinkRepository;
+import com.jujeol.drink.domain.repository.ReviewRepository;
 import com.jujeol.drink.exception.CreateReviewLimitException;
 import com.jujeol.drink.exception.NotExistReviewInDrinkException;
 import com.jujeol.drink.exception.NotFoundDrinkException;
