+++ backend/src/main/java/com/jujeol/drink/domain/Drink.java
@@ -1,5 +1,7 @@
 package com.jujeol.drink.domain;
 
+import java.util.ArrayList;
+import java.util.List;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
